% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_name.R
\name{abr_to_entidad}
\alias{abr_to_entidad}
\title{Nombre Comun completo entidad}
\usage{
abr_to_entidad(x)
}
\arguments{
\item{entidad}{Nombre de una entidad}
}
\value{
un vector de caracteres de tamaño 3 de entidades de México no
ambiguas
}
\description{
Transforma las abreviaciones más comunes de entidades de México en nombres
comunes completos. Ej: "COA" en "Coahuila". Evita nombres extensos como
"Coahuila de Zaragoza" y prefiere las apelaciones simples. Elimina caracteres
especiales, acentos y puntiacion antes de hacer la transformacion. No ve
Mayúsculas ni minúsculas. Si la entidad no se encuentra regresa NA.
}
