% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflactar_ppto.R
\name{deflactar_tp}
\alias{deflactar_tp}
\title{Deflactar montos con deflactor de Transparencia Presupuestaria}
\usage{
deflactar_tp(monto, year_monto, year_out)
}
\arguments{
\item{monto}{monto que se quiere deflactar}

\item{year_monto}{año de origen del monto}

\item{year_out}{año del precio al que se quiere deflactar}
}
\value{
regresa un vector de los montos deflactados
}
\description{
Deflactar montos con deflactor de Transparencia Presupuestaria
}
