% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_name.R
\name{entidad_to_abr}
\alias{entidad_to_abr}
\title{Abrevia una entidad}
\usage{
entidad_to_abr(entidad)
}
\arguments{
\item{entidad}{Nombre de una entidad}
}
\value{
un vector de caracteres de tamaño 3 de entidades de México no
ambiguas
}
\description{
Transforma un nombre propio de entidad Mexicana en abreviaciones no ambiguas.
Remueve todo caracter especial o acentuado y luego intenta. Si la entidad no
se encuentra regresa NA. Toda referencia a la Republica Federal o Nacion o
Nacional se transformará en "NAC". Oaxaca es un caso especial: cualquier
mencion de "oaxaca" como match de regex en el nombre se identificará con
"OAX".
}
