% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_tidy.R
\name{group_and_ponder_by}
\alias{group_and_ponder_by}
\title{Group by and Ponder}
\usage{
group_and_ponder_by(.tabla, .variable_a_pond, .ponderador, ...)
}
\arguments{
\item{.tabla}{data.frame o tabla a ponderar}

\item{.variable_a_pond}{variable que se va a ponderar en factor o String}

\item{.ponderador}{variable de probabilidades}

\item{...}{variables extra de agrupación tipo group_by anteriores a .variable_a_pond}
}
\value{
Regresa una bd con proporciones ponderadas, por grupo además de
observaciones sin ponderar (v_obs) y ponderadas (v_tot)
}
\description{
Esta función toma una base de datos y la transforma a una muestra de tidy con
.variable_a_pond como probabilidades. Funciona dentro de Tidy solamente en el
pipe (%>%).
}
\details{
Si se quiere utilizar fuera referirse a sym(!!String)
}
