% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup_row.R
\name{ungroup.rowwise_df}
\alias{ungroup.rowwise_df}
\title{Desagrupar por fila en caso de haber usado rowwise}
\usage{
\method{ungroup}{rowwise_df}(x)
}
\arguments{
\item{x}{dataframe agrupado por fila}
}
\value{
dataframe sin agrupación
}
\description{
Desagrupar por fila en caso de haber usado rowwise
}
