% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_objects.R
\name{unrowwise_df}
\alias{unrowwise_df}
\title{Desagrupar por fila en caso de haber usado rowwise}
\usage{
unrowwise_df(data)
}
\arguments{
\item{data}{dataframe agrupado por fila}
}
\value{
dataframe sin agrupación
}
\description{
Desagrupar por fila en caso de haber usado rowwise
}
